+++ src/Core/GlobalSearh/Config/Apps/appsSearchMapper.ts
@@ -1,11 +1,6 @@
 import { tokens } from '@equinor/eds-tokens';
+import { NavigateFunction } from 'react-router';
 import { SearchItem, SearchResult } from '../../Service/SearchApi';
-
-function navigate(uri: string) {
-    const url = new URL(window.location.href);
-    url.pathname = uri;
-    window.history.pushState({}, '', url);
-}
 
 export const appsSearchMapper = (data?: SearchItem[]): SearchResult | undefined => {
     return data && data.length > 0
@@ -13,7 +8,7 @@ export const appsSearchMapper = (data?: SearchItem[]): SearchResult | undefined 
               type: 'apps',
               title: 'Applications',
               color: tokens.colors.interactive.success__resting.rgba,
-              action: (id: string, item: SearchItem) => {
+              action: (id: string, item: SearchItem, navigate: NavigateFunction) => {
                   if (item.uri) {
                       window.open(item.uri, '_blank');
                       return;
